requirements.txt
flask
flask-cors
streamlit
scikit-learn
PyPDF2
python-docx
sentence-transformers
ctransformers
torch
gensim
nltk
spacy

# LLM Local (Opcional) - Respostas mais naturais e humanas
# Descomente a linha abaixo para ativar LLM local
# llama-cpp-python>=0.2.0